pyproject.toml
[project]
name = "docker-compose-psql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
  { name = "Masato Ota", email = "ohtam@nec.com" },
]
requires-python = ">=3.12"
dependencies = [
  "psycopg2-binary>=2.9.10",
  "python-dotenv>=1.1.0",
]

[project.scripts]
docker-compose-psql = "docker_compose_psql:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
